+++ pkg/analysis_server/test/lsp/server_abstract.dart
@@ -650,10 +650,10 @@ mixin LspAnalysisServerTestMixin implements ClientCapabilitiesHelperMixin {
     });
   }
 
-  String applyTextEdit(
-      String content, Either2<TextEdit, AnnotatedTextEdit> change) {
+  String applyTextEdit(String content,
+      Either3<SnippetTextEdit, AnnotatedTextEdit, TextEdit> change) {
     // Both sites of the union can cast to TextEdit.
-    final edit = change.map((e) => e, (e) => e);
+    final edit = change.map((e) => e, (e) => e, (e) => e);
     final startPos = edit.range.start;
     final endPos = edit.range.end;
     final lineInfo = LineInfo.fromContent(content);
@@ -714,8 +714,8 @@ mixin LspAnalysisServerTestMixin implements ClientCapabilitiesHelperMixin {
       );
 
     for (final change in sortedChanges) {
-      newContent = applyTextEdit(
-          newContent, Either2<TextEdit, AnnotatedTextEdit>.t1(change));
+      newContent = applyTextEdit(newContent,
+          Either3<SnippetTextEdit, AnnotatedTextEdit, TextEdit>.t3(change));
     }
 
     return newContent;
@@ -1217,6 +1217,7 @@ mixin LspAnalysisServerTestMixin implements ClientCapabilitiesHelperMixin {
     TextDocumentClientCapabilities textDocumentCapabilities,
     ClientCapabilitiesWorkspace workspaceCapabilities,
     ClientCapabilitiesWindow windowCapabilities,
+    Map<String, Object> experimentalCapabilities,
     Map<String, Object> initializationOptions,
     bool throwOnFailure = true,
     bool allowEmptyRootUri = false,
@@ -1225,6 +1226,7 @@ mixin LspAnalysisServerTestMixin implements ClientCapabilitiesHelperMixin {
       workspace: workspaceCapabilities,
       textDocument: textDocumentCapabilities,
       window: windowCapabilities,
+      experimental: experimentalCapabilities,
     );
 
     // Handle any standard incoming requests that aren't test-specific, for example