+++ pkg/compiler/lib/src/js_backend/constant_emitter.dart
@@ -18,7 +18,6 @@ import '../js_backend/string_reference.dart'
 import '../js_emitter/code_emitter_task.dart';
 import '../js_model/type_recipe.dart' show TypeExpressionRecipe;
 import '../options.dart';
-import 'field_analysis.dart' show JFieldAnalysis;
 import 'namer.dart';
 import 'runtime_types_new.dart' show RecipeEncoder;
 import 'runtime_types_resolution.dart';