+++ pkg/compiler/test/sourcemaps/helpers/sourcemap_helper.dart
@@ -22,7 +22,6 @@ import 'package:compiler/src/js/js_source_mapping.dart';
 import 'package:compiler/src/js_model/js_strategy.dart';
 import 'package:compiler/src/source_file_provider.dart';
 import '../../helpers/memory_compiler.dart';
-import '../../helpers/output_collector.dart';
 
 class SourceFileSink implements OutputSink {
   final String filename;