+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.3.0-dev
+* Added `Expression.inConstantContext` to API.
+* Updated documentation comments for some getters that don't return `null`.
+
 ## 1.2.0
 * Deprecated all setters in API of AST. Use `parseString()` instead.
 * `AnalysisSession.getErrors()` does not return `null`, check its `state`.